Added new functionality for Incr and bug fixes
Added support for substring search based increment position and end of target increment position.
Fixed errors with acquiring output folder.
Moved parser error check to after parsing.
Updated DOCS.

+++ XScrape/src/CfgParser.h
@@ -19,8 +19,8 @@ struct _Data
 	struct Incr
 	{
 		Incr() : pos(0), length(0), ammount(0) {};
-		Incr(size_t pos, size_t len, size_t ammount) : pos(pos), length(len), ammount(ammount) {};
-		size_t pos;
+		Incr(long long pos, size_t len, size_t ammount) : pos(pos), length(len), ammount(ammount) {};
+		long long pos;
 		size_t length;
 		size_t ammount;
 	};
